requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL in production
redis==5.0.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Gmail API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Background tasks
celery==5.3.4

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2 